Refactored:To Retrieve All Employees Who Have Joined In A Particular Data Range From The Payroll Service Database

+++ src/com/bridgelabz/EmployeePayrollData.java
@@ -6,11 +6,10 @@ public class EmployeePayrollData {
 	private int id;
 	private String name;
 	private String gender;
-	private int salary;
+	private double salary;
 	private Date start_date;
-	
 
-	public EmployeePayrollData(int id, String name, String gender, int salary, Date start_date) {
+	public EmployeePayrollData(int id, String name, String gender, double salary, Date start_date) {
 		this.id = id;
 		this.name = name;
 		this.gender = gender;
@@ -28,4 +27,5 @@ public class EmployeePayrollData {
 				", start_date=" + start_date +
 				'}';
 	}
+
 }
